requirements.txt
kafka-python==0.9.5
requests==2.9.0
websocket-client==0.34.0
ws4py=0.3.5
